write keys into policy (requires latest version of policy)
