+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "zplc"
-version = "0.9.1"
+version = "0.9.2"
 edition = "2024"
 
 [dependencies]
@@ -18,4 +18,4 @@ ring = "0.17.8"
 serde = { version = "1.0.217", features = ["derive"] }
 thiserror = "2.0.8"
 toml = "0.9.7"
-zpr-policy = { git = "https://github.com/org-zpr/zpr-policy-rs.git", tag = "v0.9.0", features = ["v1", "v2"]}
+zpr-policy = { git = "https://github.com/org-zpr/zpr-policy-rs.git", tag = "v0.9.1", features = ["v1", "v2"]}
